Try cmake xcode generator for ios targets
