+++ quickjs/native/CMakeLists.txt
@@ -17,10 +17,6 @@ endif ()
 option(BUILD_WITH_JNI "Build with the JNI bridge" ON)
 
 function(configure_jni include_sub_dir)
-    if (NOT BUILD_WITH_JNI)
-        return()
-    endif ()
-
     if (NOT DEFINED PLATFORM_JAVA_HOME)
         message(FATAL_ERROR "'PLATFORM_JAVA_HOME' is no set!")
     endif ()
@@ -39,16 +35,24 @@ set(all_sources "")
 # Cross compiling
 if (TARGET_PLATFORM STREQUAL "linux_x64")
     set(CMAKE_TOOLCHAIN_FILE "${CURRENT_DIR}/cmake/zig-toolchain-linux_x64.cmake")
-    configure_jni("linux")
+    if (BUILD_WITH_JNI)
+        configure_jni("linux")
+    endif ()
 elseif (TARGET_PLATFORM STREQUAL "linux_aarch64")
     set(CMAKE_TOOLCHAIN_FILE "${CURRENT_DIR}/cmake/zig-toolchain-linux_aarch64.cmake")
-    configure_jni("linux")
+    if (BUILD_WITH_JNI)
+        configure_jni("linux")
+    endif ()
 elseif (TARGET_PLATFORM STREQUAL "macos_x64")
     set(CMAKE_TOOLCHAIN_FILE "${CURRENT_DIR}/cmake/zig-toolchain-macos_x64.cmake")
-    configure_jni("darwin")
+    if (BUILD_WITH_JNI)
+        configure_jni("darwin")
+    endif ()
 elseif (TARGET_PLATFORM STREQUAL "macos_aarch64")
     set(CMAKE_TOOLCHAIN_FILE "${CURRENT_DIR}/cmake/zig-toolchain-macos_aarch64.cmake")
-    configure_jni("darwin")
+    if (BUILD_WITH_JNI)
+        configure_jni("darwin")
+    endif ()
 elseif (TARGET_PLATFORM STREQUAL "ios_x64")
     set(CMAKE_TOOLCHAIN_FILE "${CURRENT_DIR}/cmake/zig-toolchain-ios_x64.cmake")
 elseif (TARGET_PLATFORM STREQUAL "ios_aarch64")
@@ -56,7 +60,9 @@ elseif (TARGET_PLATFORM STREQUAL "ios_aarch64")
 elseif (TARGET_PLATFORM STREQUAL "ios_aarch64_simulator")
     set(CMAKE_TOOLCHAIN_FILE "${CURRENT_DIR}/cmake/zig-toolchain-ios_aarch64_simulator.cmake")
 elseif (TARGET_PLATFORM STREQUAL "windows_x64")
-    configure_jni("win32")
+    if (BUILD_WITH_JNI)
+        configure_jni("win32")
+    endif ()
     set(CMAKE_TOOLCHAIN_FILE "${CURRENT_DIR}/cmake/zig-toolchain-windows_x64.cmake")
     include_directories("winpthreads/include")
     if (BUILD_WITH_JNI)
